Extract host button actions to a separate module.
